# packages/contracts/openapi.v1.yaml
openapi: 3.1.0
info:
  title: The Penthouse API
  version: 1.0.0-alpha
  description: Private social platform API. All protected endpoints require a Bearer access token.
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: string
              description: Human-readable error message
            - type: object
              description: Zod field-level validation error (fieldErrors + formErrors)
...
    UserRole:
      type: string
      enum: [admin, member]
    UserStatus:
      type: string
      enum: [active, removed, banned]
...
    Message:
      type: object
      required: [id, chatId, senderId, content, type, createdAt]
      properties:
        id:
          type: string
          format: uuid
        chatId:
          type: string
          format: uuid
        senderId:
          type: string
          format: uuid
        senderUsername:
          type: string
          description: Immutable login identifier. May be omitted in future member-facing clients.
        senderDisplayName:
          type: string
        senderAvatarUrl:
          type: string
          nullable: true
        content:
          type: string
        type:
          type: string
          enum: [text, image, video, gif, file]
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        clientMessageId:
          type: string
          description: Client-supplied idempotency key (optional on received messages)
...
    MemberSummary:
      type: object
      required: [id, username, displayName, avatarUrl]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
    MemberDetail:
      allOf:
        - $ref: '#/components/schemas/MemberSummary'
        - type: object
          required: [bio]
          properties:
            bio:
              type: string
              nullable: true
    AdminMemberSummary:
      allOf:
        - $ref: '#/components/schemas/MemberDetail'
        - type: object
          required: [role, status, mustChangePassword, createdAt]
          properties:
            role:
              $ref: '#/components/schemas/UserRole'
            status:
              $ref: '#/components/schemas/UserStatus'
            mustChangePassword:
              type: boolean
            createdAt:
              type: string
              format: date-time
    AdminInviteResponse:
      type: object
      required: [code, uses, maxUses, createdAt]
      properties:
        code:
          type: string
        uses:
          type: integer
          minimum: 0
        maxUses:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
    AdminTempPasswordResponse:
      type: object
      required: [userId, username, temporaryPassword]
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
        temporaryPassword:
          type: string
    AdminMessage:
      allOf:
        - $ref: '#/components/schemas/Message'
        - type: object
          required: [senderStatus, hidden]
          properties:
            senderStatus:
              $ref: '#/components/schemas/UserStatus'
            hidden:
              type: boolean
paths:
...
  /api/v1/members:
    get:
      summary: Search active members
      operationId: listMembers
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          description: Case-insensitive search across username and display name
          schema:
            type: string
      responses:
        '200':
          description: Active member directory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Password change required or account unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/admin/invite:
    get:
      summary: Get the current master invite code
      operationId: adminGetInvite
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current master invite
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminInviteResponse'
        '403':
          description: Admin only, password change required, or account unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/invite/rotate:
    post:
      summary: Rotate the master invite code
      operationId: adminRotateInvite
      security:
        - bearerAuth: []
      responses:
        '200':
          description: New master invite code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminInviteResponse'
        '403':
          description: Admin only, password change required, or account unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/members:
    get:
      summary: Search all members, including removed and banned accounts
      operationId: adminListMembers
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Full member directory for admin review
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminMemberSummary'
        '403':
          description: Admin only, password change required, or account unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/members/{memberId}/remove:
    post:
      summary: Remove a member account
      operationId: adminRemoveMember
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Member removed, refresh tokens revoked, and memberships cleared
        '403':
          description: Admin only, password change required, or account unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Cannot remove self or another admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/members/{memberId}/ban:
    post:
      summary: Ban a member account
      operationId: adminBanMember
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Member banned, refresh tokens revoked, and memberships cleared
        '403':
          description: Admin only, password change required, or account unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Cannot ban self or another admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/members/{memberId}/temp-password:
    post:
      summary: Issue a temporary password and force password change
      operationId: adminIssueTempPassword
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Temporary password issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTempPasswordResponse'
        '403':
          description: Admin only, password change required, or account unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Cannot modify self or another admin, or target account is not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/chats/{chatId}/messages:
    get:
      summary: Audit chat history including hidden messages
      operationId: adminListChatMessages
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
      responses:
        '200':
          description: Full audit-visible chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminMessage'
        '403':
          description: Admin only, password change required, or account unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...